+++ api/integrations/pagseguro/pagseguro.config.js
