+++ api/services/services.delivery.js
@@ -15,9 +15,9 @@ const emailUtils = require('../utils/email/email.index')
 const emailUpdateSolicitation = require('../utils/email/email.update_payment')
 const deliveryMapper = require('../mappers/mappers.delivery')
 
-const config = require('../utils/utils.correios')
-const { calcBox } = require('../utils/helpers/helpers.calcBox')
-const ErrorGeneric = require('../utils/errors/erros.generic-error')
+const config = require('../integrations/correios/correios.config')
+const { calcBox } = require('../integrations/correios/correios.index')
+const ErrorGeneric = require('../exceptions/erros.generic-error')
 
 const listByIdDeliveryService = async (id) => {
   try {
