+++ api/services/services.product.js
@@ -1,7 +1,7 @@
 const { ObjectId } = require('mongodb')
 const { product, rating, client, category } = require('../models/models.index')
 const productMapper = require('../mappers/mappers.product')
-const ErrorGeneric = require('../utils/errors/erros.generic-error')
+const ErrorGeneric = require('../exceptions/erros.generic-error')
 
 const getSort = (sortType) => {
   switch (sortType) {
