+++ api/services/services.user.js
@@ -3,10 +3,10 @@ const cryptography = require('../utils/utils.cryptography')
 const emailUtils = require('../utils/email/email.index')
 const { Email } = require('../utils/email/email.recovery')
 const userMapper = require('../mappers/mappers.user')
-const ErrorGeneric = require('../utils/errors/erros.generic-error')
-const ErrorNotAuthenticatedUser = require('../utils/errors/errors.user-not-authenticated')
-const ErrorNotAuthorizedUser = require('../utils/errors/errors.user-not-authorized')
-const ErrorBusinessRule = require('../utils/errors/errors.business-rule')
+const ErrorGeneric = require('../exceptions/erros.generic-error')
+const ErrorNotAuthenticatedUser = require('../exceptions/errors.user-not-authenticated')
+const ErrorNotAuthorizedUser = require('../exceptions/errors.user-not-authorized')
+const ErrorBusinessRule = require('../exceptions/errors.business-rule')
 const profile = require('../utils/utils.rules')
 
 const userIsValidService = async (email, password) => {
